# openapi.yaml
openapi: 3.1.0
info:
  title: Samokoder SaaS API
  version: 1.0.0
  description: |
    Полнофункциональный API для платформы Samokoder - AI-powered генератор кода.
    
    ## Аутентификация
    API использует JWT токены с Bearer схемой аутентификации. Токены передаются в заголовке:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Безопасность
    - JWT токены с ротацией (access: 15 минут, refresh: 7 дней)
    - Rate limiting на критических эндпоинтах
    - Защита от brute-force атак
    - CORS с белым списком доменов
    - Валидация всех входных данных
    
    ## Версионирование
    API версионируется через префикс `/v1`. Мы гарантируем обратную совместимость в рамках мажорной версии.
    
  contact:
    name: Samokoder Support
    email: support@samokoder.io
  license:
    name: Proprietary
servers:
  - url: https://api.samokoder.io
    description: Production сервер
  - url: https://staging-api.samokoder.io
    description: Staging сервер
  - url: http://localhost:8000
    description: Локальная разработка
...
paths:
...
  /v1/auth/register:
    post:
      tags:
        - auth
      summary: Регистрация нового пользователя
      description: |
        Создает нового пользователя с валидацией пароля согласно ASVS 2.1.1:
        - Минимум 8 символов
        - Минимум 1 заглавная буква
        - Минимум 1 строчная буква
        - Минимум 1 цифра
        - Минимум 1 специальный символ
        - Не должен быть в списке распространенных паролей
        
        **Реализация:** `api/routers/auth.py:141-168`
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Пользователь уже существует или невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /v1/auth/login:
    post:
      tags:
        - auth
      summary: Вход в систему
      description: |
        Аутентификация пользователя с email и паролем.
        
        **Защита:**
        - Rate limiting: 5 запросов в минуту
        - Блокировка после 5 неудачных попыток на 15 минут
        - Аудит лог всех попыток входа
        - HttpOnly cookies для токенов
        
        **Реализация:** `api/routers/auth.py:171-257`
      operationId: login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  format: email
                  description: Email пользователя
                password:
                  type: string
                  format: password
                  description: Пароль пользователя
              required:
                - username
                - password
      responses:
        '200':
          description: Успешная аутентификация
          headers:
            Set-Cookie:
              description: HttpOnly cookies с access_token и refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Невалидные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          description: Слишком много попыток входа - аккаунт заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-rate-limit:
        limit: 5
        window: 60
  /v1/auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление access токена
      description: |
        Выдает новый access токен на основе refresh токена.
        
        **Реализация:** `api/routers/auth.py:260-288`
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Токен успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          description: Невалидный refresh токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-rate-limit:
        limit: 10
        window: 60
  /v1/auth/logout:
    post:
      tags:
        - auth
      summary: Выход из системы
      description: |
        Отзывает текущий access токен (добавляет jti в черный список).
        
        **Реализация:** `api/routers/auth.py:291-322`
      operationId: logout
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully logged out"
      security:
        - BearerAuth: []
  /v1/auth/me:
    get:
      tags:
        - auth
      summary: Получить профиль текущего пользователя
      description: |
        Возвращает базовую информацию о текущем пользователе.
        
        **Реализация:** `api/routers/auth.py:325-334`
      operationId: getCurrentUser
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/auth/github:
    get:
      tags:
        - auth
      summary: Инициировать GitHub OAuth
      description: |
        Перенаправляет на GitHub для OAuth авторизации.
        
        **Реализация:** `api/routers/auth.py:337-349`
      operationId: initiateGitHubAuth
      responses:
        '302':
          description: Перенаправление на GitHub
        '400':
          description: GitHub OAuth не настроен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/auth/github/callback:
    get:
      tags:
        - auth
      summary: GitHub OAuth callback
      description: |
        Обрабатывает callback от GitHub OAuth и сохраняет токен.
        
        **Реализация:** `api/routers/auth.py:352-393`
      operationId: handleGitHubCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Код авторизации от GitHub
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: JWT токен с user_id
      responses:
        '302':
          description: Перенаправление на страницу настроек
        '400':
          description: Невалидный state или ошибка получения токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен в формате: Bearer <token>
        
        Получить токен можно через `/v1/auth/login` или `/v1/auth/register`
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email пользователя
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: |
            Пароль должен содержать:
            - Минимум 8 символов
            - Минимум 1 заглавную букву
            - Минимум 1 строчную букву
            - Минимум 1 цифру
            - Минимум 1 специальный символ
          example: "SecurePass123!"
...
    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user_id
        - email
      properties:
        access_token:
          type: string
          description: JWT access токен
        refresh_token:
          type: string
          description: JWT refresh токен
        token_type:
          type: string
          enum: [bearer]
          default: bearer
        expires_in:
          type: integer
          description: Время жизни access токена в секундах
          example: 900
        user_id:
          type: integer
          description: ID пользователя
        email:
          type: string
          format: email
          description: Email пользователя
    TokenRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh токен
    TokenRefreshResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: Новый JWT access токен
        token_type:
          type: string
          enum: [bearer]
          default: bearer
        expires_in:
          type: integer
          description: Время жизни токена в секундах
          example: 900
    UserResponse:
      type: object
      required:
        - id
        - email
        - tier
        - created_at
        - projects_count
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        tier:
          type: string
          enum: [free, pro, enterprise]
        created_at:
          type: string
          format: date-time
        projects_count:
          type: integer
          minimum: 0

    # Project schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Описание ошибки
        error_code:
          type: string
          description: Код ошибки (опционально)
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
security:
  - BearerAuth: []
...